generator/requirements.txt
psycopg2-binary==2.9.9
numpy==2.1.1
